// types/doctor.types.ts
export interface Doctor {
    id: number;
    title: string;
    firstname: string;
    lastname: string;
    working_place: string;
    available_days: string;
    bmdc_code: string;
    job_designation: string;
    booking_phone: string;
    gender: string;
    degree_name: string;
    consultancy_fee: string;
    mobile: string;
    provide_service: string;
    about: string;
    email: string;
    speciality: string;
    password: string;
    starting_pratice: string;
    achievement: string;
    user_id: number;
    created_at: string;
    updated_at: string;
}